+++ lib/const.ts
@@ -1,1 +1,2 @@
 export const PAGE_BY = 8;
+export const BLOG_HANDLE = "Journal";