+++ lib/shopify/queries/blog.ts
@@ -1,0 +1,37 @@
+export const BLOGS_QUERY = `#graphql
+query Blog(
+  $language: LanguageCode
+  $blogHandle: String!
+  $pageBy: Int!
+  $cursor: String
+) @inContext(language: $language) {
+  blog(handle: $blogHandle) {
+    title
+    seo {
+      title
+      description
+    }
+    articles(first: $pageBy, after: $cursor) {
+      edges {
+        node {
+          author: authorV2 {
+            name
+          }
+          contentHtml
+          handle
+          id
+          image {
+            id
+            altText
+            url
+            width
+            height
+          }
+          publishedAt
+          title
+        }
+      }
+    }
+  }
+}
+`;